batman/requirements.txt
# Core dependencies
jsonschema>=4.17.3
pyyaml>=6.0.1
prometheus-client>=0.17.1
psutil>=5.9.5

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
isort>=5.12.0

# Documentation
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0 